refactor: ecowatt
